[refactoring] Move update_url_params from functions (deleted) to helpers.urls
